refactor(kb): split KBService into functional modules
Replace the monolithic KBService class with per-operation functions under
articles/, knowledge-base/, and internal/ helpers.
